Add most missing html elements Still need some patient characteristic, TODOs, and unit fixes
